.openclaw/workspace/projects/Axe-Capital/step1-data-foundation/pyproject.toml
[project]
name = "axe-coo"
version = "0.1.0"
description = "Axe Capital Step 1: COO data agent connectors + normalization"
requires-python = ">=3.11"
dependencies = [
  "ddgs>=9.5.0",
  "httpx>=0.27",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "praw>=7.7",
  "yfinance>=0.2.50",
]

[project.scripts]
axe-coo-bundle = "axe_coo.cli:main"
axe-step2-analyze = "axe_coo.step2_cli:main"
axe-step3-decide = "axe_coo.step3_cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
